Only detach from Catalyst thread if one was actually started.

+++ t/catalyst_startup.pm
@@ -66,7 +66,7 @@ sub start_catalyst {
 }
 
 sub stop_catalyst {
-    $catalyst_thread->detach();
+    $catalyst_thread->detach() if $catalyst_thread;
 }
 
 sub wait_for_catalyst {
